+++ src/instructions/u.rs
@@ -29,6 +29,6 @@ pub fn parse_u_inst(raw_word: InstructionWord) -> Format {
     }
 }
 
-pub fn execute_u_type(op: &UOp, rd: usize, imm: usize, register: &mut RegisterFile) -> Result<ExecutionSignal, String> {
+pub fn execute_u_type(op: &UOp, rd: usize, imm: i32, register: &mut RegisterFile) -> Result<ExecutionSignal, String> {
     Ok(ExecutionSignal::Continue)
 }
